refactor(app-log): move policy into daemon ownership
