+++ src/daemon/__tests__/managed-session-artifact-path.test.ts
@@ -2,7 +2,7 @@ import fs from 'node:fs';
 import os from 'node:os';
 import path from 'node:path';
 import { afterEach, expect, test } from 'vitest';
-import { requireManagedSessionArtifactPath } from './managed-session-artifact-path.ts';
+import { requireManagedSessionArtifactPath } from '../managed-session-artifact-path.ts';
 
 const roots: string[] = [];
