+++ src/daemon/managed-session-artifact-path.ts
