+++ src/platform-runtime-app-log-output.ts
@@ -1,8 +1,8 @@
 import type { AppLogOutputSink } from '@agent-device/contracts/app-log-runtime';
 import fs from 'node:fs';
 import { openVerifiedFileForAppend, openVerifiedFileForRead } from '@agent-device/host-kit/file';
-import { ensureAppLogPath } from './utils/app-log-files.ts';
-import { requireManagedSessionArtifactPath } from './utils/managed-session-artifact-path.ts';
+import { ensureAppLogPath } from './daemon/app-log-files.ts';
+import { requireManagedSessionArtifactPath } from './daemon/managed-session-artifact-path.ts';
 
 export async function openAppLogOutput(
   sessionsDir: string,
